moved update function out of main to allow for easy mass-updating

+++ source/draw.c
@@ -1,15 +1,15 @@
 #include "draw.h"
 
-void drawMenu(char ** names, u8 * state, u32 selected_entry)
+void drawMenu(char ** names, u8 * state, u8 selected_entry)
 {
 	for (u8 i = 0; i <= 40; i++) {
 		printf("\x1b[0;40;37m\x1b[5;%uH=", (4+i));
 		printf("\x1b[0;40;37m\x1b[24;%uH=", (4+i));
 	}
 	
-	for(u32 i = 0; i < 18; i++) {
+	for(u8 i = 0; i < 18; i++) {
 		char format[64];
-		sprintf(format, "\x1b[%lu;4H", (i+6));
+		sprintf(format, "\x1b[%u;4H", (i+6));
 		
 		if (i == selected_entry )
 		{
