+++ source/draw.h
@@ -1,3 +1,3 @@
 #include "basic.h"
 
-void drawMenu(char ** names, u8 * state, u32 selected_entry);
+void drawMenu(char ** names, u8 * state, u8 selected_entry);
